+++ frameworks/Java/quarkus/base/src/main/resources/fortunes.mustache
