+++ frameworks/Java/quarkus/benchmark_config.json
@@ -5,9 +5,6 @@
       "default": {
         "json_url": "/json",
         "plaintext_url": "/plaintext",
-        "db_url": "/db",
-        "query_url": "/queries?queries=",
-        "fortune_url": "/fortunes",
         "port": 8080,
         "approach": "Realistic",
         "classification": "fullstack",
@@ -20,7 +17,52 @@
         "webserver": "None",
         "os": "Linux",
         "database_os": "Linux",
-        "display_name": "Quarkus",
+        "display_name": "quarkus",
+        "notes": "",
+        "versus": "Netty"
+      }
+    },
+    {
+      "hibernate": {
+        "db_url": "/hibernate/db",
+        "query_url": "/hibernate/queries?queries=",
+        "fortune_url": "/hibernate/fortunes",
+        "port": 8080,
+        "approach": "Realistic",
+        "classification": "fullstack",
+        "database": "Postgres",
+        "framework": "Quarkus",
+        "language": "Java",
+        "flavor": "None",
+        "orm": "full",
+        "platform": "JAX-RS",
+        "webserver": "None",
+        "os": "Linux",
+        "database_os": "Linux",
+        "display_name": "quarkus-hibernate",
+        "notes": "",
+        "versus": "Netty"
+      }
+    },
+    {
+      "pgclient": {
+        "db_url": "/pgclient/db",
+        "query_url": "/pgclient/queries?queries=",
+        "update_url": "/pgclient/updates?queries=",
+        "fortune_url": "/pgclient/fortunes",
+        "port": 8080,
+        "approach": "Realistic",
+        "classification": "fullstack",
+        "database": "Postgres",
+        "framework": "Quarkus",
+        "language": "Java",
+        "flavor": "None",
+        "orm": "Micro",
+        "platform": "JAX-RS",
+        "webserver": "None",
+        "os": "Linux",
+        "database_os": "Linux",
+        "display_name": "quarkus-pgclient",
         "notes": "",
         "versus": "Netty"
       }
