+++ frameworks/Java/quarkus/hibernate/src/main/resources/application.properties
@@ -1,0 +1,16 @@
+quarkus.datasource.url=jdbc:postgresql://tfb-database:5432/hello_world
+quarkus.datasource.driver=org.postgresql.Driver
+quarkus.datasource.username=benchmarkdbuser
+quarkus.datasource.password=benchmarkdbpass
+quarkus.datasource.max-size=64
+quarkus.datasource.min-size=16
+
+quarkus.log.console.enable=true
+quarkus.log.console.level=DEBUG
+quarkus.log.file.enable=false
+quarkus.log.level=INFO
+quarkus.log.category."org.hibernate".level=INFO
+
+quarkus.hibernate-orm.log.sql=false
+
+quarkus.thread-pool.queue-size=2147483647
