+++ frameworks/Java/quarkus/pgclient/src/main/resources/application.properties
@@ -1,0 +1,9 @@
+quarkus.datasource.url=vertx-reactive:postgresql://tfb-database:5432/hello_world
+quarkus.datasource.username=benchmarkdbuser
+quarkus.datasource.password=benchmarkdbpass
+quarkus.datasource.max-size=64
+quarkus.log.console.enable=true
+quarkus.log.console.level=INFO
+quarkus.log.file.enable=false
+quarkus.log.level=INFO
+quarkus.thread-pool.queue-size=2147483647
