--- frameworks/Java/quarkus/src/main/java/io/quarkus/benchmark/filter/ServerHeaderFilter.java
@@ -1,15 +1,0 @@
-package io.quarkus.benchmark.filter;
-
-import javax.ws.rs.container.ContainerRequestContext;
-import javax.ws.rs.container.ContainerResponseContext;
-import javax.ws.rs.container.ContainerResponseFilter;
-import javax.ws.rs.ext.Provider;
-
-@Provider
-public class ServerHeaderFilter implements ContainerResponseFilter {
-
-    @Override
-    public void filter(ContainerRequestContext requestContext, ContainerResponseContext responseContext) {
-        responseContext.getHeaders().add("Server", "Quarkus");
-    }
-}
