--- frameworks/Java/quarkus/src/main/java/io/quarkus/benchmark/repository/FortuneRepository.java
@@ -1,25 +1,0 @@
-package io.quarkus.benchmark.repository;
-
-import java.util.List;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
-import javax.persistence.EntityManager;
-import javax.persistence.Query;
-import javax.transaction.Transactional;
-
-import io.quarkus.benchmark.model.Fortune;
-import io.quarkus.benchmark.model.World;
-
-@ApplicationScoped
-public class FortuneRepository {
-
-    @Inject
-    EntityManager em;
-
-    @Transactional
-    public List<Fortune> findAll() {
-        Query query = em.createQuery("SELECT f FROM Fortune f");
-        return query.getResultList();
-    }
-}
